backend/requirements.txt
azure-cognitiveservices-speech==1.20.0
bert-extractive-summarizer==0.10.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cymem==2.0.6
en-core-web-sm==3.2.0
ffmpeg-python==0.2.0
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
future==0.18.2
huggingface-hub==0.4.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.22.2
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyee==9.0.4
pyparsing==3.0.7
python-dotenv==0.19.2
python-ffmpeg==1.0.13
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.6.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.10.2
tqdm==4.62.3
transformers==4.16.2
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
Werkzeug==2.0.3
